リファクタリング (#455)
* テスト修正

* リファクタリング: 座席調整ロジックを別ファイルに分離

システムの座席調整ロジックを system/core/seat_adjustment.go に抽出し、system.go から削除。これにより、コードの可読性と保守性を向上。主な変更点は以下の通り:

- AdjustMaxSeats メソッドを新しいファイルに移動
- adjustGeneralSeats と adjustMemberSeats メソッドを同様に移動
- システムコアの関数を小さな関数に分割し、関心の分離を実現

* リファクタリング: WorkspaceAppクラスの実装を分割

システムの主要な機能を持つWorkspaceAppクラスを複数のファイルに分割し、コードの構造と可読性を改善。主な変更点は以下の通り:

- core/workspaceapp/ ディレクトリに以下のファイルを追加:
  - type_workspace_app.go: WorkspaceAppの型定義
  - workspace_app.go: メインの実装
  - workspace_app_utils.go: ユーティリティメソッド
  - validator.go: コマンドバリデーション関連のメソッド
  - batch.go: バッチ処理関連のメソッド
  - seat_adjustment.go: 座席調整関連のメソッド

- 既存のシステムコードを新しいWorkspaceAppクラスに移行
- コードの関心の分離を実現
- テストケースを更新

* refatoパッケージ構造の改修・ファイル分割

* refactor: テストファイル分割
