+++ system/aws-lambda/transfer_collection_history_bigquery/main.go
@@ -1,13 +1,13 @@
 package main
 
 import (
+	"app.modules/core/workspaceapp"
 	"context"
 	"log/slog"
 
 	"app.modules/core/utils"
 
 	"app.modules/aws-lambda/lambdautils"
-	"app.modules/core"
 	"github.com/aws/aws-lambda-go/lambda"
 )
 
@@ -24,7 +24,7 @@ func TransferCollectionHistoryBigquery() (TransferCollectionHistoryBigqueryRespo
 	if err != nil {
 		return TransferCollectionHistoryBigqueryResponse{}, nil
 	}
-	sys, err := core.NewSystem(ctx, false, clientOption)
+	sys, err := workspaceapp.NewSystem(ctx, false, clientOption)
 	if err != nil {
 		return TransferCollectionHistoryBigqueryResponse{}, nil
 	}
