+++ system/aws-lambda/youtube_organize_database/main.go
@@ -1,11 +1,11 @@
 package main
 
 import (
+	"app.modules/core/workspaceapp"
 	"context"
 	"log/slog"
 
 	"app.modules/aws-lambda/lambdautils"
-	"app.modules/core"
 	"app.modules/core/utils"
 	"github.com/aws/aws-lambda-go/lambda"
 )
@@ -23,7 +23,7 @@ func OrganizeDatabase() (OrganizeDatabaseResponse, error) {
 	if err != nil {
 		return OrganizeDatabaseResponse{}, nil
 	}
-	system, err := core.NewSystem(ctx, false, clientOption)
+	system, err := workspaceapp.NewSystem(ctx, false, clientOption)
 	if err != nil {
 		return OrganizeDatabaseResponse{}, nil
 	}
